+++ properties/Connection/AllowToStartTwoQueriesInParallel.php
@@ -3,23 +3,24 @@ declare(strict_types = 1);
 
 namespace Properties\Formal\AccessLayer\Connection;
 
-use Formal\AccessLayer\Query\SQL;
+use Formal\AccessLayer\{
+    Query\SQL,
+    Connection,
+};
 use Innmind\BlackBox\{
     Property,
     Set,
+    Runner\Assert,
 };
-use PHPUnit\Framework\Assert;
 
+/**
+ * @implements Property<Connection>
+ */
 final class AllowToStartTwoQueriesInParallel implements Property
 {
     public static function any(): Set
     {
-        return Set\Property::of(self::class);
-    }
-
-    public function name(): string
-    {
-        return 'Allow to start 2 queries in parallel';
+        return Set\Elements::of(new self);
     }
 
     public function applicableTo(object $connection): bool
@@ -27,14 +28,14 @@ final class AllowToStartTwoQueriesInParallel implements Property
         return true;
     }
 
-    public function ensureHeldBy(object $connection): object
+    public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $result1 = $connection(SQL::of('show tables'));
         $result2 = $connection(SQL::of('show tables'));
 
         // by using any() we only do a partial iteration over the results
-        Assert::assertTrue($result1->any(static fn() => true));
-        Assert::assertTrue($result2->any(static fn() => true));
+        $assert->true($result1->any(static fn() => true));
+        $assert->true($result2->any(static fn() => true));
 
         return $connection;
     }