+++ properties/Connection/AnInvalidLazyQueryMustThrow.php
@@ -6,23 +6,22 @@ namespace Properties\Formal\AccessLayer\Connection;
 use Formal\AccessLayer\{
     Query\SQL,
     Exception\QueryFailed,
+    Connection,
 };
 use Innmind\BlackBox\{
     Property,
     Set,
+    Runner\Assert,
 };
-use PHPUnit\Framework\Assert;
 
+/**
+ * @implements Property<Connection>
+ */
 final class AnInvalidLazyQueryMustThrow implements Property
 {
     public static function any(): Set
     {
-        return Set\Property::of(self::class);
-    }
-
-    public function name(): string
-    {
-        return 'An invalid lazy query must throw';
+        return Set\Elements::of(new self);
     }
 
     public function applicableTo(object $connection): bool
@@ -30,7 +29,7 @@ final class AnInvalidLazyQueryMustThrow implements Property
         return true;
     }
 
-    public function ensureHeldBy(object $connection): object
+    public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $query = SQL::onDemand('INSERT');
         $result = $connection($query);
@@ -38,11 +37,11 @@ final class AnInvalidLazyQueryMustThrow implements Property
         try {
             // throw only now because the force the execution of the sequence
             $result->toList();
-            Assert::fail('it should throw an exception');
+            $assert->fail('it should throw an exception');
         } catch (QueryFailed $e) {
-            Assert::assertSame($query, $e->query());
-            Assert::assertIsInt($e->code());
-            Assert::assertIsString($e->message());
+            $assert->same($query, $e->query());
+            $assert->number($e->code())->int();
+            $assert->string($e->message());
         }
 
         return $connection;