+++ properties/Connection/AnInvalidLazySelectMustThrow.php
@@ -7,23 +7,22 @@ use Formal\AccessLayer\{
     Query\Select,
     Table\Name,
     Exception\QueryFailed,
+    Connection,
 };
 use Innmind\BlackBox\{
     Property,
     Set,
+    Runner\Assert,
 };
-use PHPUnit\Framework\Assert;
 
+/**
+ * @implements Property<Connection>
+ */
 final class AnInvalidLazySelectMustThrow implements Property
 {
     public static function any(): Set
     {
-        return Set\Property::of(self::class);
-    }
-
-    public function name(): string
-    {
-        return 'An invalid lazy SELECT must throw';
+        return Set\Elements::of(new self);
     }
 
     public function applicableTo(object $connection): bool
@@ -31,7 +30,7 @@ final class AnInvalidLazySelectMustThrow implements Property
         return true;
     }
 
-    public function ensureHeldBy(object $connection): object
+    public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $query = Select::onDemand(new Name('unknown'));
         $result = $connection($query);
@@ -39,11 +38,11 @@ final class AnInvalidLazySelectMustThrow implements Property
         try {
             // throw only now because the force the execution of the sequence
             $result->toList();
-            Assert::fail('it should throw an exception');
+            $assert->fail('it should throw an exception');
         } catch (QueryFailed $e) {
-            Assert::assertSame($query, $e->query());
-            Assert::assertIsInt($e->code());
-            Assert::assertIsString($e->message());
+            $assert->same($query, $e->query());
+            $assert->number($e->code())->int();
+            $assert->string($e->message());
         }
 
         return $connection;