+++ properties/Connection/SelectAliasedColumns.php
@@ -6,8 +6,9 @@ namespace Properties\Formal\AccessLayer\Connection;
 use Formal\AccessLayer\{
     Query\SQL,
     Query\Parameter,
-    Query\StartTransaction,
-    Query\Commit,
+    Query\Select,
+    Table\Name,
+    Table\Column,
     Connection,
 };
 use Innmind\BlackBox\{
@@ -19,7 +20,7 @@ use Innmind\BlackBox\{
 /**
  * @implements Property<Connection>
  */
-final class ContentInsertedAfterStartOfTransactionIsAccessible implements Property
+final class SelectAliasedColumns implements Property
 {
     private string $uuid;
     private string $username;
@@ -49,49 +50,41 @@ final class ContentInsertedAfterStartOfTransactionIsAccessible implements Proper
 
     public function ensureHeldBy(Assert $assert, object $connection): object
     {
-        $connection(new StartTransaction);
-
         $insert = SQL::of('INSERT INTO `test` VALUES (?, ?, ?);')
             ->with(Parameter::of($this->uuid))
             ->with(Parameter::of($this->username))
             ->with(Parameter::of($this->number));
         $connection($insert);
 
-        $rows = $connection(SQL::of("SELECT * FROM `test` WHERE `id` = '{$this->uuid}'"));
+        $select = Select::from($table = Name::of('test')->as('_test'))->columns(
+            Column\Name::of('id')->in($table)->as('test_id'),
+            Column\Name::of('id')->as('test_id_bis'),
+        );
+        $rows = $connection($select);
 
-        $assert->count(1, $rows);
-        $assert->same(
-            $this->uuid,
-            $rows
-                ->first()
-                ->flatMap(static fn($row) => $row->column('id'))
-                ->match(
-                    static fn($id) => $id,
-                    static fn() => null,
-                ),
-        );
-        $assert->same(
-            $this->username,
-            $rows
-                ->first()
-                ->flatMap(static fn($row) => $row->column('username'))
-                ->match(
-                    static fn($username) => $username,
-                    static fn() => null,
-                ),
-        );
-        $assert->same(
-            $this->number,
-            $rows
-                ->first()
-                ->flatMap(static fn($row) => $row->column('registerNumber'))
-                ->match(
-                    static fn($registerNumber) => $registerNumber,
-                    static fn() => null,
-                ),
-        );
-
-        $connection(new Commit);
+        $assert
+            ->number($rows->size())
+            ->greaterThanOrEqual(1);
+        $assert->true($rows->first()->match(
+            static fn($row) => $row->contains('test_id'),
+            static fn() => null,
+        ));
+        $assert->true($rows->first()->match(
+            static fn($row) => $row->contains('test_id_bis'),
+            static fn() => null,
+        ));
+        $assert->false($rows->first()->match(
+            static fn($row) => $row->contains('id'),
+            static fn() => null,
+        ));
+        $assert->false($rows->first()->match(
+            static fn($row) => $row->contains('username'),
+            static fn() => null,
+        ));
+        $assert->false($rows->first()->match(
+            static fn($row) => $row->contains('registerNumber'),
+            static fn() => null,
+        ));
 
         return $connection;
     }