+++ properties/Connection/SelectEverything.php
@@ -8,13 +8,17 @@ use Formal\AccessLayer\{
     Query\Parameter,
     Query\Select,
     Table\Name,
+    Connection,
 };
 use Innmind\BlackBox\{
     Property,
     Set,
+    Runner\Assert,
 };
-use PHPUnit\Framework\Assert;
 
+/**
+ * @implements Property<Connection>
+ */
 final class SelectEverything implements Property
 {
     private string $uuid;
@@ -30,17 +34,12 @@ final class SelectEverything implements Property
 
     public static function any(): Set
     {
-        return Set\Property::of(
-            self::class,
+        return Set\Composite::immutable(
+            static fn(...$args) => new self(...$args),
             Set\Uuid::any(),
             Set\Strings::madeOf(Set\Chars::ascii())->between(0, 255),
             Set\Integers::any(),
         );
-    }
-
-    public function name(): string
-    {
-        return 'Select everything';
     }
 
     public function applicableTo(object $connection): bool
@@ -48,7 +47,7 @@ final class SelectEverything implements Property
         return true;
     }
 
-    public function ensureHeldBy(object $connection): object
+    public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $insert = SQL::of('INSERT INTO `test` VALUES (?, ?, ?);')
             ->with(Parameter::of($this->uuid))
@@ -58,16 +57,18 @@ final class SelectEverything implements Property
 
         $rows = $connection(Select::from(new Name('test')));
 
-        Assert::assertGreaterThanOrEqual(1, $rows->size());
-        Assert::assertTrue($rows->first()->match(
+        $assert
+            ->number($rows->size())
+            ->greaterThanOrEqual(1);
+        $assert->true($rows->first()->match(
             static fn($row) => $row->contains('id'),
             static fn() => null,
         ));
-        Assert::assertTrue($rows->first()->match(
+        $assert->true($rows->first()->match(
             static fn($row) => $row->contains('username'),
             static fn() => null,
         ));
-        Assert::assertTrue($rows->first()->match(
+        $assert->true($rows->first()->match(
             static fn($row) => $row->contains('registerNumber'),
             static fn() => null,
         ));