+++ properties/Connection/SelectWhere.php
@@ -8,6 +8,7 @@ use Formal\AccessLayer\{
     Query\Parameter,
     Query\Select,
     Table\Name,
+    Connection,
 };
 use Innmind\Specification\{
     Comparator,
@@ -17,9 +18,12 @@ use Innmind\Specification\{
 use Innmind\BlackBox\{
     Property,
     Set,
+    Runner\Assert,
 };
-use PHPUnit\Framework\Assert;
 
+/**
+ * @implements Property<Connection>
+ */
 final class SelectWhere implements Property
 {
     private string $uuid;
@@ -35,17 +39,12 @@ final class SelectWhere implements Property
 
     public static function any(): Set
     {
-        return Set\Property::of(
-            self::class,
+        return Set\Composite::immutable(
+            static fn(...$args) => new self(...$args),
             Set\Uuid::any(),
             Set\Strings::madeOf(Set\Chars::ascii())->between(0, 255),
             Set\Integers::any(),
         );
-    }
-
-    public function name(): string
-    {
-        return 'Select where';
     }
 
     public function applicableTo(object $connection): bool
@@ -53,7 +52,7 @@ final class SelectWhere implements Property
         return true;
     }
 
-    public function ensureHeldBy(object $connection): object
+    public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $insert = SQL::of('INSERT INTO `test` VALUES (?, ?, ?);')
             ->with(Parameter::of($this->uuid))
@@ -89,8 +88,8 @@ final class SelectWhere implements Property
         });
         $rows = $connection($select);
 
-        Assert::assertCount(1, $rows);
-        Assert::assertSame(
+        $assert->count(1, $rows);
+        $assert->same(
             $this->uuid,
             $rows
                 ->first()