+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 2.2.0 - 2023-07-30
+
+### Added
+
+- `Formal\AccessLayer\Table\Name::of()`
+- `Formal\AccessLayer\Table\Name::as()`
+- `Formal\AccessLayer\Table\Column::of()`
+- `Formal\AccessLayer\Table\Column\Name::of()`
+- `Formal\AccessLayer\Table\Column\Name::in()`
+- `Formal\AccessLayer\Table\Column\Name::as()`
+- `Formal\AccessLayer\Table\Name\Aliased`
+- `Formal\AccessLayer\Table\Column\Name\Namespaced`
+- `Formal\AccessLayer\Table\Column\Name\Aliased`
+
+### Changed
+
+- Require `innmind/black-box` `5`
+
+### Removed
+
+- Support for PHP `8.1`
+
 ## 2.1.0 - 2023-07-07
 
 ### Added